Add irf/edisp_kernel.py
