+++ gammapy/irf/__init__.py
@@ -6,6 +6,7 @@ from .background import *
 from .effective_area import *
 from .energy_dispersion import *
 from .edisp_map import *
+from .edisp_kernel import *
 from .io import *
 from .irf_reduce import *
 from .irf_stack import *
